+++ R/tar_meta_upload.R
@@ -27,10 +27,33 @@
 #' })
 #' }
 tar_meta_upload <- function(
+  meta = TRUE,
+  progress = TRUE,
+  process = TRUE,
+  crew = TRUE,
   verbose = TRUE,
+  strict = FALSE,
   script = targets::tar_config_get("script"),
   store = targets::tar_config_get("store")
 ) {
+  tar_assert_lgl(meta)
+  tar_assert_scalar(meta)
+  tar_assert_none_na(meta)
+  tar_assert_lgl(progress)
+  tar_assert_scalar(progress)
+  tar_assert_none_na(progress)
+  tar_assert_lgl(process)
+  tar_assert_scalar(process)
+  tar_assert_none_na(process)
+  tar_assert_lgl(crew)
+  tar_assert_scalar(crew)
+  tar_assert_none_na(crew)
+  tar_assert_lgl(verbose)
+  tar_assert_scalar(verbose)
+  tar_assert_none_na(verbose)
+  tar_assert_lgl(strict)
+  tar_assert_scalar(strict)
+  tar_assert_none_na(strict)
   tar_assert_script(script)
   tar_assert_store(store)
   options <- tar_option_script(script = script)
@@ -42,9 +65,29 @@ tar_meta_upload <- function(
   })
   tar_options$set_repository_meta(options$repository_meta)
   tar_options$set_resources(options$resources)
-  database_meta(path_store = store)$upload(verbose = verbose)
-  database_progress(path_store = store)$upload(verbose = verbose)
-  database_process(path_store = store)$upload(verbose = verbose)
-  database_crew(path_store = store)$upload(verbose = verbose)
+  if (meta) {
+    database_meta(path_store = store)$nice_upload(
+      verbose = verbose,
+      strict = strict
+    )
+  }
+  if (progress) {
+    database_progress(path_store = store)$nice_upload(
+      verbose = verbose,
+      strict = strict
+    )
+  }
+  if (process) {
+    database_process(path_store = store)$nice_upload(
+      verbose = verbose,
+      strict = strict
+    )
+  }
+  if (crew) {
+    database_crew(path_store = store)$nice_upload(
+      verbose = verbose,
+      strict = strict
+    )
+  }
   invisible()
 }