+++ tests/aws/test-aws_meta.R
@@ -191,7 +191,7 @@ tar_test("AWS tar_meta_upload()", {
       )
     )
   }
-  tar_meta_delete(which = "cloud")
+  tar_meta_delete(delete = "cloud")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       aws_s3_exists(
@@ -252,7 +252,7 @@ tar_test("AWS tar_meta_download()", {
       file.exists(file.path(path_meta_dir(path_store_default()), file))
     )
   }
-  tar_meta_delete(which = "local")
+  tar_meta_delete(delete = "local")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       file.exists(file.path(path_meta_dir(path_store_default()), file))
@@ -309,7 +309,7 @@ tar_test("AWS tar_meta_sync() upload", {
       )
     )
   }
-  tar_meta_delete(which = "cloud")
+  tar_meta_delete(delete = "cloud")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       aws_s3_exists(
@@ -370,7 +370,7 @@ tar_test("AWS tar_meta_sync() download", {
       file.exists(file.path(path_meta_dir(path_store_default()), file))
     )
   }
-  tar_meta_delete(which = "local")
+  tar_meta_delete(delete = "local")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       file.exists(file.path(path_meta_dir(path_store_default()), file))
@@ -380,6 +380,78 @@ tar_test("AWS tar_meta_sync() download", {
   for (file in c("meta", "process", "progress", "crew")) {
     expect_true(
       file.exists(file.path(path_meta_dir(path_store_default()), file))
+    )
+  }
+})
+
+tar_test("AWS tar_meta_sync() download graceful failure", {
+  skip_if_no_aws()
+  s3 <- paws.storage::s3()
+  bucket_name <- random_bucket_name()
+  s3$create_bucket(Bucket = bucket_name)
+  on.exit(aws_s3_delete_bucket(bucket_name))
+  code <- substitute({
+    library(targets)
+    tar_option_set(
+      repository = "aws",
+      resources = tar_resources(
+        aws = tar_resources_aws(
+          bucket = bucket_name,
+          prefix = "_targets"
+        )
+      )
+    )
+    list(tar_target(a, 1L))
+  }, env = list(bucket_name = bucket_name))
+  do.call(tar_script, list(code = code))
+  dir.create(path_store_default())
+  expect_message(
+    tar_meta_download(verbose = TRUE, strict = FALSE),
+    class = "tar_condition_run"
+  )
+  expect_silent(
+    tar_meta_download(verbose = FALSE, strict = FALSE)
+  )
+  for (verbose in c(TRUE, FALSE)) {
+    expect_error(
+      tar_meta_download(verbose = verbose, strict = TRUE),
+      class = "tar_condition_run"
+    )
+  }
+})
+
+tar_test("AWS tar_meta_sync() upload graceful failure", {
+  skip_if_no_aws()
+  s3 <- paws.storage::s3()
+  bucket_name <- random_bucket_name()
+  s3$create_bucket(Bucket = bucket_name)
+  on.exit(aws_s3_delete_bucket(bucket_name))
+  code <- substitute({
+    library(targets)
+    tar_option_set(
+      repository = "aws",
+      resources = tar_resources(
+        aws = tar_resources_aws(
+          bucket = bucket_name,
+          prefix = "_targets"
+        )
+      )
+    )
+    list(tar_target(a, 1L))
+  }, env = list(bucket_name = bucket_name))
+  do.call(tar_script, list(code = code))
+  dir.create(path_store_default())
+  expect_message(
+    tar_meta_upload(verbose = TRUE, strict = FALSE),
+    class = "tar_condition_run"
+  )
+  expect_silent(
+    tar_meta_upload(verbose = FALSE, strict = FALSE)
+  )
+  for (verbose in c(TRUE, FALSE)) {
+    expect_error(
+      tar_meta_upload(verbose = verbose, strict = TRUE),
+      class = "tar_condition_run"
     )
   }
 })