+++ tests/gcp/test-gcp_meta.R
@@ -194,7 +194,7 @@ tar_test("gcp tar_meta_upload()", {
       )
     )
   }
-  tar_meta_delete(which = "cloud")
+  tar_meta_delete(delete = "cloud")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       gcp_gcs_exists(
@@ -256,7 +256,7 @@ tar_test("gcp tar_meta_download()", {
       file.exists(file.path(path_meta_dir(path_store_default()), file))
     )
   }
-  tar_meta_delete(which = "local")
+  tar_meta_delete(delete = "local")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       file.exists(file.path(path_meta_dir(path_store_default()), file))
@@ -314,7 +314,7 @@ tar_test("gcp tar_meta_sync() upload", {
       )
     )
   }
-  tar_meta_delete(which = "cloud")
+  tar_meta_delete(delete = "cloud")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       gcp_gcs_exists(
@@ -376,7 +376,7 @@ tar_test("gcp tar_meta_sync() download", {
       file.exists(file.path(path_meta_dir(path_store_default()), file))
     )
   }
-  tar_meta_delete(which = "local")
+  tar_meta_delete(delete = "local")
   for (file in c("meta", "process", "progress", "crew")) {
     expect_false(
       file.exists(file.path(path_meta_dir(path_store_default()), file))
@@ -386,6 +386,82 @@ tar_test("gcp tar_meta_sync() download", {
   for (file in c("meta", "process", "progress", "crew")) {
     expect_true(
       file.exists(file.path(path_meta_dir(path_store_default()), file))
+    )
+  }
+})
+
+tar_test("GCP tar_meta_sync() download graceful failure", {
+  skip_if_no_gcp()
+  bucket_name <- random_bucket_name()
+  project <- Sys.getenv("GCE_DEFAULT_PROJECT_ID")
+  gcp_gcs_auth(max_tries = 5)
+  googleCloudStorageR::gcs_create_bucket(bucket_name, projectId = project)
+  on.exit(gcp_gcs_delete_bucket(bucket_name))
+  code <- substitute({
+    library(targets)
+    tar_option_set(
+      format = "rds",
+      repository = "gcp",
+      resources = tar_resources(
+        gcp = tar_resources_gcp(
+          bucket = bucket_name,
+          prefix = "_targets"
+        )
+      )
+    )
+    list(tar_target(a, 1L))
+  }, env = list(bucket_name = bucket_name))
+  do.call(tar_script, list(code = code))
+  dir.create(path_store_default())
+  expect_message(
+    tar_meta_download(verbose = TRUE, strict = FALSE),
+    class = "tar_condition_run"
+  )
+  expect_silent(
+    tar_meta_download(verbose = FALSE, strict = FALSE)
+  )
+  for (verbose in c(TRUE, FALSE)) {
+    expect_error(
+      tar_meta_download(verbose = verbose, strict = TRUE),
+      class = "tar_condition_run"
+    )
+  }
+})
+
+tar_test("GCP tar_meta_sync() upload graceful failure", {
+  skip_if_no_gcp()
+  bucket_name <- random_bucket_name()
+  project <- Sys.getenv("GCE_DEFAULT_PROJECT_ID")
+  gcp_gcs_auth(max_tries = 5)
+  googleCloudStorageR::gcs_create_bucket(bucket_name, projectId = project)
+  on.exit(gcp_gcs_delete_bucket(bucket_name))
+  code <- substitute({
+    library(targets)
+    tar_option_set(
+      format = "rds",
+      repository = "gcp",
+      resources = tar_resources(
+        gcp = tar_resources_gcp(
+          bucket = bucket_name,
+          prefix = "_targets"
+        )
+      )
+    )
+    list(tar_target(a, 1L))
+  }, env = list(bucket_name = bucket_name))
+  do.call(tar_script, list(code = code))
+  dir.create(path_store_default())
+  expect_message(
+    tar_meta_upload(verbose = TRUE, strict = FALSE),
+    class = "tar_condition_run"
+  )
+  expect_silent(
+    tar_meta_upload(verbose = FALSE, strict = FALSE)
+  )
+  for (verbose in c(TRUE, FALSE)) {
+    expect_error(
+      tar_meta_upload(verbose = verbose, strict = TRUE),
+      class = "tar_condition_run"
     )
   }
 })